Renamed fuelSavings util
