+++ src/reducers/fuelSavingsReducer.js
@@ -1,5 +1,5 @@
 import {SAVE_FUEL_SAVINGS, CALCULATE_FUEL_SAVINGS} from '../constants/actionTypes';
-import {necessaryDataIsProvidedToCalculateSavings, calculateSavings} from '../utils/fuelSavingsCalculator';
+import {necessaryDataIsProvidedToCalculateSavings, calculateSavings} from '../utils/fuelSavings';
 import objectAssign from 'object-assign';
 import initialState from './initialState';
