+++ src/store/store.spec.js
@@ -3,7 +3,7 @@ import * as ActionTypes from '../constants/actionTypes';
 import MockDate from 'mockdate';
 import configureStore from './configureStore';
 
-import {calculateSavings} from '../utils/fuelSavingsCalculator';
+import {calculateSavings} from '../utils/fuelSavings';
 import {getFormattedDateTime} from '../utils/dates';
 
 describe('Store', () => {
